new_quiz_app/requirements.txt
blinker==1.6.3
cachelib==0.13.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.7
cryptography==43.0.3
dnspython==2.7.0
Flask==2.3.3
Flask-PyMongo==2.3.0
Flask-Session==0.5.0
idna==3.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
msal==1.25.0
pycparser==2.22
PyJWT==2.10.1
pymongo==4.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.32.5
six==1.17.0
urllib3==2.5.0
Werkzeug==2.3.7
